fix(#239): Delete game with existing plays
- Fix Navbar for attendee-check-in (small fix)
- Add two tests for deleting play to win and library
- Fixed multiple issues with deleting play to win games
- Fixed library games removing foreign key constraints

closes: #239

+++ apps/admin/app/api/library/[id]/route.ts
@@ -78,6 +78,20 @@ export async function GET(
  *                message: string
  */
 export async function DELETE(request: NextRequest, { params }: { params: { id: string } }) {
+  const libraryItem = await prisma.libraryItem.findFirst({
+    where: { id: Number(params.id)}
+  })
+
+  if (libraryItem === null || libraryItem === undefined)
+    return NextResponse.json({ message: `Library item with ${params.id} not found` }, { status: 404 });
+
+
+  await prisma.libraryCheckoutEvent.deleteMany({
+    where: {
+      libraryCopyId: Number(params.id)
+    }
+  })
+
   await prisma.libraryItem.delete({
     where: { id: Number(params.id) },
   });
